Fix train on windows

+++ mmdet/core/bbox/assigners/assign_result.py
@@ -57,7 +57,7 @@ class AssignResult(object):
             [self.max_overlaps.new_ones(len(gt_labels)), self.max_overlaps])
 
         if self.labels is not None:
-            self.labels = torch.cat([gt_labels, self.labels])
+            self.labels = torch.cat([gt_labels.long(), self.labels])
 
     def __nice__(self):
         """
